apps/api/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
pgvector==0.2.4

# Task Queue
celery==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# AI/ML
# Note: These will be installed in the specific adapters
# google-generativeai
# openai
# anthropic
sentence-transformers==2.3.0
transformers==4.37.0
torch==2.1.0
spacy==3.7.2
scikit-learn==1.4.0
numpy==1.26.0

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
langdetect==1.0.9
pypdf==4.0.0
python-magic==0.4.27

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
faker==22.0.0
